refactor: split Student Project Detail scene's protocols into separated folder and files
