+++ Pharus/Scenes/Student Project Detail/Protocols/StudentProjectDetailFlow.swift
@@ -1,0 +1,14 @@
+//
+//  StudentProjectDetailFlow.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol StudentProjectDetailFlow {
+    func showProjectRules()
+    func showSendFileView()
+    func showMentorReview()
+}
