+++ Pharus/Scenes/Student Project Detail/StudentProjectDetailPresenter.swift
@@ -7,13 +7,6 @@
 
 import UIKit
 
-protocol StudentProjectsDetailPresenterProtocol {
-    func showProjectRules()
-    func showUploadFileView()
-    func toggleTaskCompletedStatus(task: TaskModel)
-    func showMentorReview()
-}
-
 class StudentProjectDetailPresenter: StudentProjectsDetailPresenterProtocol {
 
     // MARK: - Properties
